+++ src/index.ts
@@ -1,2 +1,3 @@
 export * from './format-date-time/format-date-time';
 export * from './date-time-difference/date-time-difference';
+export * from './index-to-name/index-to-name';